+++ src/iterator-for-combination.cpp
@@ -1,32 +1,31 @@
+#include <algorithm>
 #include <string>
 #include <vector>
 
 using namespace std;
 
 class CombinationIterator {
-private:
+ private:
   vector<string> combinations;
   int idx;
 
-  void dfs(const string &characters, int idx, int curr, int n_set, int cnt) {
-    if (idx == characters.size()) {
+  void dfs(const string& characters, int idx, int curr, int n_set, int cnt) {
+    if (idx == (int)characters.size()) {
       if (n_set == cnt) {
         string ans = "";
         for (size_t i = 0; i < characters.size(); ++i) {
-          if (curr & (1 << i))
-            ans += characters[characters.size() - i - 1];
+          if (curr & (1 << i)) ans += characters[characters.size() - i - 1];
         }
         reverse(ans.begin(), ans.end());
         combinations.push_back(ans);
       }
       return;
     }
-    if (n_set < cnt)
-      dfs(characters, idx + 1, (curr << 1) + 1, n_set + 1, cnt);
+    if (n_set < cnt) dfs(characters, idx + 1, (curr << 1) + 1, n_set + 1, cnt);
     dfs(characters, idx + 1, curr << 1, n_set, cnt);
   }
 
-public:
+ public:
   CombinationIterator(string characters, int combinationLength) {
     dfs(characters, 0, 0, 0, combinationLength);
     idx = 0;
@@ -34,5 +33,5 @@ public:
 
   string next() { return combinations[idx++]; }
 
-  bool hasNext() { return idx < combinations.size(); }
+  bool hasNext() { return idx < (int)combinations.size(); }
 };
