+++ src/subsets.cpp
@@ -3,13 +3,12 @@
 using namespace std;
 
 class Solution {
-private:
-  void aux(vector<int> &nums, vector<bool> &include, int N, vector<vector<int>> &res) {
-    if (N == nums.size()) {
+ private:
+  void aux(vector<int>& nums, vector<bool>& include, int N, vector<vector<int>>& res) {
+    if (N == (int)nums.size()) {
       vector<int> n;
       for (int i = 0; i < N; i++) {
-        if (include[i])
-          n.push_back(nums[i]);
+        if (include[i]) n.push_back(nums[i]);
       }
       res.push_back(n);
     } else {
@@ -20,8 +19,8 @@ private:
     }
   }
 
-public:
-  vector<vector<int>> subsets(vector<int> &nums) {
+ public:
+  vector<vector<int>> subsets(vector<int>& nums) {
     vector<bool> include;
     include.assign(nums.size(), false);
     vector<vector<int>> res;
