+++ src/the-k-th-lexicographical-string-of-all-happy-strings-of-length-n.cpp
@@ -1,33 +1,32 @@
+#include <algorithm>
 #include <string>
 #include <vector>
 
 using namespace std;
 
 class Solution {
-private:
+ private:
   vector<string> strings;
 
   void gen(int n, string curr = "") {
-    if (curr.size() == n) {
+    if ((int)curr.size() == n) {
       strings.push_back(curr);
       return;
     }
 
     for (char c = 'a'; c <= 'c'; ++c) {
-      if (!curr.empty() && c == curr.back())
-        continue;
+      if (!curr.empty() && c == curr.back()) continue;
 
       gen(n, curr + c);
     }
   }
 
-public:
+ public:
   string getHappyString(int n, int k) {
     strings.clear();
     gen(n);
     sort(strings.begin(), strings.end());
-    if (k > strings.size())
-      return "";
+    if (k > (int)strings.size()) return "";
     return strings[k - 1];
   }
 };
