+++ src/top-k-frequent-elements.cpp
@@ -5,11 +5,10 @@
 using namespace std;
 
 class Solution {
-public:
-  vector<int> topKFrequent(vector<int> &nums, int k) {
+ public:
+  vector<int> topKFrequent(vector<int>& nums, int k) {
     unordered_map<int, int> f;
-    for (auto e : nums)
-      f[e]++;
+    for (auto e : nums) f[e]++;
 
     priority_queue<pair<int, int>> pq;
     for (auto e : f) {
@@ -17,7 +16,7 @@ public:
     }
 
     vector<int> res;
-    while (res.size() != k) {
+    while ((int)res.size() != k) {
       auto curr = pq.top();
       pq.pop();
       res.push_back(curr.second);
